[refactor] controller 메서드 단위로 매핑
