+++ src/main/java/com/ceos22/cgv_clone/domain/movie/entity/Movie.java
@@ -1,7 +1,6 @@
 package com.ceos22.cgv_clone.domain.movie.entity;
 
 import com.ceos22.cgv_clone.domain.common.entity.BaseEntity;
-import com.ceos22.cgv_clone.domain.movie.AgeRating;
 import jakarta.persistence.*;
 import lombok.*;
