+++ src/main/java/com/ceos22/cgv_clone/domain/reservation/controller/TicketController.java
@@ -10,7 +10,6 @@ import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
-@RequestMapping("/tickets")
 @RequiredArgsConstructor
 @SecurityRequirement(name = "Bearer Authentication") // JWT 인증 필요
 public class TicketController {
@@ -18,13 +17,13 @@ public class TicketController {
     private final TicketService ticketService;
 
     // 예매하기
-    @PostMapping("/reserve")
+    @PostMapping("/tickets/reserve")
     public TicketResponseDto reserveTicket(@AuthenticationPrincipal UserDetails userDetails, @RequestBody TicketRequestDto request) {
         return ticketService.reserveTicket(userDetails.getUsername(), request);
     }
 
     // 예매 취소하기
-    @DeleteMapping("/cancel/{ticketId}")
+    @DeleteMapping("/tickets/cancel/{ticketId}")
     public String cancelTicket(@AuthenticationPrincipal UserDetails userDetails, @PathVariable Long ticketId) {
         ticketService.cancelTicket(userDetails.getUsername(), ticketId);
         return "예매가 취소되었습니다.";
