+++ src/main/java/com/ceos22/cgv_clone/domain/reservation/entity/ReservationStatus.java
@@ -1,4 +1,4 @@
-package com.ceos22.cgv_clone.domain.reservation;
+package com.ceos22.cgv_clone.domain.reservation.entity;
 
 public enum ReservationStatus {
     AVAILABLE,
