+++ src/main/java/com/ceos22/cgv_clone/domain/shop/controller/OrderController.java
@@ -10,18 +10,17 @@ import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
-@RequestMapping("/orders")
 @RequiredArgsConstructor
 @SecurityRequirement(name = "Bearer Authentication") // JWT 인증 필요
 public class OrderController {
     private final OrderService orderService;
 
-    @PostMapping
+    @PostMapping("/orders")
     public OrderResponseDto createOrder(@AuthenticationPrincipal UserDetails userDetails, @RequestBody OrderRequestDto orderRequestDto) {
         return orderService.createOrder(userDetails.getUsername(), orderRequestDto);
     }
 
-    @GetMapping("/{orderId}")
+    @GetMapping("/orders/{orderId}")
     public OrderResponseDto getOrder(@AuthenticationPrincipal UserDetails userDetails, @PathVariable Long orderId) {
         return orderService.getOrder(userDetails.getUsername(), orderId);
     }
