+++ src/main/java/com/ceos22/cgv_clone/domain/theater/entity/ScreenType.java
@@ -1,4 +1,4 @@
-package com.ceos22.cgv_clone.domain.theater;
+package com.ceos22.cgv_clone.domain.theater.entity;
 
 public enum ScreenType {
     NORMAL("일반관"),
