+++ src/main/java/com/ceos22/cgv_clone/domain/user/controller/AuthController.java
@@ -9,21 +9,20 @@ import org.springframework.web.bind.annotation.*;
 
 
 @RestController
-@RequestMapping("/auth")
 @RequiredArgsConstructor
 public class AuthController {
 
     private final AuthService authService;
 
     // 회원가입
-    @PostMapping("/signup")
+    @PostMapping("/auth/signup")
     public String signup(@RequestBody SignupRequestDto request) {
         authService.signup(request.getUsername(), request.getPassword());
         return "회원가입 성공";
     }
 
     // 로그인
-    @PostMapping("/login")
+    @PostMapping("/auth/login")
     public LoginResponseDto login(@RequestBody LoginRequestDto request) {
         String token = authService.login(request.getUsername(), request.getPassword());
         return new LoginResponseDto(token, "Bearer");
